Refactored the plinko module into independent files.
